shared/requirements.txt
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
asyncpg==0.28.0
envparse>=0.2.0
python-dotenv>=1.0.0
structlog==23.1.0
motor==3.3.1
aio-pika==9.2.3
pika==1.3.2
sentry-sdk==1.34.0
redis==5.1.0a1
yookassa~=3.0.0
celery==5.3.4
